+++ fmpz_mat/mul_blas.c
@@ -97,7 +97,7 @@ static void fmpz_multi_mod_uint32_stride(
     }
     else
     {
-        _fmpz_multi_mod_run(A, C->mod_P, input, -1, CT->T);
+        _fmpz_multi_mod_precomp(A, C->mod_P, input, -1, CT->T);
     }
 
     offsets = C->mod_offsets;
